Option Fix bed respawn location not resetting
As it happens, players on a server with a huge game world somehow need to travel long distances.
From Minecraft version 1.16 (≈June 2020) to version 1.17.1(≈October 2021) in the game there was a bug SPIGOT-5988 which did not reset the respawn point of the player after death, if his bed was blocked.
After some manipulations, it turned out to be a simple mechanism that allowed the player to get to the spavn by blocking his bed with a shalker before death, and after the second death to be near his bed again.
This bug persisted from the Spigot server to Paper and all its forks until October 2021, after which it was fixed by the Spigot development team.
Attempts by players to reach out to Spigot to allow them to disable the patch that fixes the SPIGOT-5988 bug have failed.

Demonstration of how the patch works:
https://www.youtube.com/watch?v=3y5SbQXzMss&feature=youtu.be